Advclass fix and cleanup (#3580)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Potential fix for double pick exploit? How? Cleanup hugbox start and end
code and some procs

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fix

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

+++ code/controllers/subsystem/role_class_handler/role_class_handler.dm
@@ -114,32 +114,26 @@ SUBSYSTEM_DEF(role_class_handler)
 	Since this is class handler related, might as well also have the class handler send itself into the params
 */
 /datum/controller/subsystem/role_class_handler/proc/finish_class_handler(mob/living/carbon/human/H, datum/job/advclass/picked_class, datum/class_select_handler/related_handler, special_session_queue)
-	if(!picked_class || !related_handler || !H) // Extreme edge case but is possible, likely href exploit or late activation
+	if(!picked_class || QDELETED(related_handler) || QDELETED(H)) // Extreme edge case but is possible, likely href exploit or late activation
 		return FALSE
+
 	if(picked_class.total_positions != -1)
 		if(picked_class.current_positions >= picked_class.total_positions)
 			related_handler.rolled_class_is_full(picked_class)
 			return FALSE
-	H.advsetup = FALSE // This is actually on a lot of shit, so its a ghetto selector protector if u need one
+
+	SSjob.AssignRole(H, picked_class, TRUE)
+	adjust_class_amount(picked_class, 1) // we are handling one guy right now.
+	H.finish_class_hugbox()
+
+	qdel(related_handler)
+
 	if(picked_class.inherit_parent_title)
 		var/datum/job/old = SSjob.GetJob(H.job)
 		if(old)
 			picked_class.title = old.title
-	SSjob.AssignRole(H, picked_class, TRUE)
+
 	SSjob.EquipRank(H, picked_class, H.client)
-	H.density = TRUE
-	H.invisibility = initial(H.invisibility)
-	H.cure_blind("advsetup")
-	var/atom/movable/screen/advsetup/GET_IT_OUT = locate() in H.hud_used.static_inventory //locate() still iterates over contents
-	if(GET_IT_OUT)
-		qdel(GET_IT_OUT)
-	// We don't track open browsers, so we need to force close them here
-	related_handler.ForceCloseMenus() // force menus closed
-
-	class_select_handlers.Remove(related_handler.linked_client.ckey)
-	qdel(related_handler)
-
-	adjust_class_amount(picked_class, 1) // we are handling one guy right now.
 
 // A dum helper to adjust the class amount, we could do it elsewhere but this will also inform any relevant class handlers open.
 /datum/controller/subsystem/role_class_handler/proc/adjust_class_amount(datum/job/advclass/target_datum, amount)
@@ -155,12 +149,8 @@ SUBSYSTEM_DEF(role_class_handler)
 			found_menu.rolled_class_is_full(target_datum) //  inform the datum of its error.
 
 /datum/controller/subsystem/role_class_handler/proc/cancel_class_handler(mob/living/carbon/human/H)
-	H.advsetup = FALSE
-	H.invisibility = 0
-	var/atom/movable/screen/advsetup/GET_IT_OUT = locate() in H.hud_used.static_inventory //locate() still iterates over contents
-	qdel(GET_IT_OUT)
-	H.cure_blind("advsetup")
+	H.finish_class_hugbox()
 
 	var/datum/class_select_handler/related_handler = class_select_handlers[H.client.ckey]
-	related_handler?.ForceCloseMenus() // force menus closed
-	qdel(related_handler)
+	if(related_handler)
+		qdel(related_handler)
